# Add button benchmark