--- a/crates/egui_demo_lib/benches/benchmark.rs
+++ b/crates/egui_demo_lib/benches/benchmark.rs
@@ -4,9 +4,16 @@ use criterion::{criterion_group, criterion_main, BatchSize, Criterion};
 
 use egui::epaint::TextShape;
 use egui::load::SizedTexture;
-use egui::{Button, TextureId, UiBuilder, Vec2};
+use egui::{Button, Id, TextureId, Ui, UiBuilder, Vec2};
 use egui_demo_lib::LOREM_IPSUM_LONG;
 use rand::Rng as _;
+
+/// Each iteration should be called in their own `Ui` with an intentional id clash,
+/// to prevent the Context from building a massive map of `WidgetRects` (which would slow the test,
+/// causing unreliable results).
+fn create_benchmark_ui(ctx: &egui::Context) -> Ui {
+    Ui::new(ctx.clone(), Id::new("clashing_id"), UiBuilder::new())
+}
 
 pub fn criterion_benchmark(c: &mut Criterion) {
     use egui::RawInput;
@@ -57,25 +64,23 @@ pub fn criterion_benchmark(c: &mut Criterion) {
     {
         let ctx = egui::Context::default();
         let _ = ctx.run(RawInput::default(), |ctx| {
-            egui::CentralPanel::default().show(ctx, |ui| {
-                c.bench_function("label &str", |b| {
-                    b.iter_batched_ref(
-                        || ui.new_child(UiBuilder::new()),
-                        |ui| {
-                            ui.label("the quick brown fox jumps over the lazy dog");
-                        },
-                        BatchSize::LargeInput,
-                    )
-                });
-                c.bench_function("label format!", |b| {
-                    b.iter_batched_ref(
-                        || ui.new_child(UiBuilder::new()),
-                        |ui| {
-                            ui.label("the quick brown fox jumps over the lazy dog".to_owned());
-                        },
-                        BatchSize::LargeInput,
-                    )
-                });
+            c.bench_function("label &str", |b| {
+                b.iter_batched_ref(
+                    || create_benchmark_ui(ctx),
+                    |ui| {
+                        ui.label("the quick brown fox jumps over the lazy dog");
+                    },
+                    BatchSize::LargeInput,
+                );
+            });
+            c.bench_function("label format!", |b| {
+                b.iter_batched_ref(
+                    || create_benchmark_ui(ctx),
+                    |ui| {
+                        ui.label("the quick brown fox jumps over the lazy dog".to_owned());
+                    },
+                    BatchSize::LargeInput,
+                );
             });
         });
     }
@@ -83,40 +88,38 @@ pub fn criterion_benchmark(c: &mut Criterion) {
     {
         let ctx = egui::Context::default();
         let _ = ctx.run(RawInput::default(), |ctx| {
-            egui::CentralPanel::default().show(ctx, |ui| {
-                let mut group = c.benchmark_group("button");
+            let mut group = c.benchmark_group("button");
 
-                // To ensure we have a valid image, let's use the font texture. The size
-                // shouldn't be important for this benchmark.
-                let image = SizedTexture::new(TextureId::default(), Vec2::splat(16.0));
+            // To ensure we have a valid image, let's use the font texture. The size
+            // shouldn't be important for this benchmark.
+            let image = SizedTexture::new(TextureId::default(), Vec2::splat(16.0));
 
-                group.bench_function("button", |b| {
-                    b.iter_batched_ref(
-                        || ui.new_child(UiBuilder::new()),
-                        |ui| {
-                            ui.add(Button::new("Hello World"));
-                        },
-                        BatchSize::LargeInput,
-                    );
-                });
-                group.bench_function("button image", |b| {
-                    b.iter_batched_ref(
-                        || ui.new_child(UiBuilder::new()),
-                        |ui| {
-                            ui.add(Button::image_and_text(image, "Hello World"));
-                        },
-                        BatchSize::LargeInput,
-                    )
-                });
-                group.bench_function("button image right text", |b| {
-                    b.iter_batched_ref(
-                        || ui.new_child(UiBuilder::new()),
-                        |ui| {
-                            ui.add(Button::image_and_text(image, "Hello World").right_text("⏵"));
-                        },
-                        BatchSize::LargeInput,
-                    )
-                });
+            group.bench_function("1_button_text", |b| {
+                b.iter_batched_ref(
+                    || create_benchmark_ui(ctx),
+                    |ui| {
+                        ui.add(Button::new("Hello World"));
+                    },
+                    BatchSize::LargeInput,
+                );
+            });
+            group.bench_function("2_button_text_image", |b| {
+                b.iter_batched_ref(
+                    || create_benchmark_ui(ctx),
+                    |ui| {
+                        ui.add(Button::image_and_text(image, "Hello World"));
+                    },
+                    BatchSize::LargeInput,
+                );
+            });
+            group.bench_function("3_button_text_image_right_text", |b| {
+                b.iter_batched_ref(
+                    || create_benchmark_ui(ctx),
+                    |ui| {
+                        ui.add(Button::image_and_text(image, "Hello World").right_text("⏵"));
+                    },
+                    BatchSize::LargeInput,
+                );
             });
         });
     }